Build: Use AM_V_GEN in Makefiles where appropriate
When generating new files as part of the Makefile the recipe is
sometimes suppressed with $(AM_V_GEN) and sometimes with `@`. We should
prefer $(AM_V_GEN), since this also prints the lines in silent mode.
This is arguably more in style with the current recipe echoing.

Before:
Generated test/data/script_tests.json.h
Now:
  GEN      test/data/script_tests.json.h

A side effect of this change is that the recipe for generating build.h
is now echoed on each make run. Arguably this makes its generation more
transparent.